+++ tests/north_project/sql/1.1/1.1-index-ddl.sql
@@ -1,0 +1,2 @@
+CREATE INDEX CONCURRENTLY north_app_author_idx ON north_app_author(id);
+CREATE INDEX CONCURRENTLY north_app_book_idx ON north_app_book(id);